preparando para linux

+++ src/main/java/py/edu/facitec/ventas/dto/PaginadorDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 @Data
 public class PaginadorDto<T> {
     private List<T> items;        // Lista de la página actual
-    private int totalItems;       // Total de elementos
+    private long totalItems;       // Total de elementos
     private int totalPages;       // Total de páginas
     private int currentPage;      // Página actual (1-indexada)
     private int itemsCount;       // Cantidad de items en la página actual
